minor: Rename `createStore` to `create-store`
